requirements.txt
Flask==2.3.2
flask-cors==5.0.1
werkzeug==3.0.1

PyPDF2==3.0.1
python-docx==1.1.2

selenium==4.29.0
webdriver-manager==4.0.2
beautifulsoup4==4.12.2

pandas==2.2.1
openpyxl==3.1.2

google-generativeai==0.8.4
python-dotenv==1.0.1
